refactor(streaming): extract segment packaging into a service
Move fMP4/TS segment serving — the CMAF rewrite dispatch, tfdt timeline
anchoring, and the per-rendition init-track cache — out of the 2162-line
StreamingController into a focused SegmentPackagingService under
streaming/services/ (the first service in the module's new services/ subfolder).

The controller's segment routes now call segmentPackaging.serve(); the segment
duration is passed in explicitly (deferring removal of the module-global
SEG_DURATION to the later PlaybackPlan work, per the plan) and the init-track
cache is encapsulated as a private field instead of a module global.

Behaviour-preserving: the streaming-lifecycle integration spec (which exercises
segment serving + tfdt anchoring) and the manifest goldens pass unchanged.
First step toward splitting the controller god-file.
