+++ backend/src/modules/streaming/streaming.module.ts
@@ -21,6 +21,7 @@ import { SubtitleBurnInService } from './subtitle-burn-in.service';
 import { ThumbnailService } from './thumbnail.service';
 import { RecommendationService } from './recommendation.service';
 import { StreamingSettingsCache } from './streaming-settings-cache.service';
+import { SegmentPackagingService } from './services/segment-packaging.service';
 import { Command } from '../scheduler/entities/command.entity';
 import { AuthModule } from '../auth/auth.module';
 import { SettingsModule } from '../settings/settings.module';
@@ -58,6 +59,7 @@ import { MarkersModule } from '../markers/markers.module';
     ThumbnailService,
     RecommendationService,
     StreamingSettingsCache,
+    SegmentPackagingService,
   ],
   exports: [
     PlaybackService,
